requirements.txt
numpy>=1.24.0
pandas>=2.0.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
PyYAML>=6.0.0
